BjarkeB

+++ API/controllers/foodController.js
@@ -1,34 +1,50 @@
 // importing food model
 import Food from "../models/food.js"
 
-// exporting get all method - gets all foods
+// gets all foods
 export const getFoods = async (req, res) => {
     try {
-        Food.find()
-        .then((foods) => {
-            // returns status 200 and all foods
-            res.status(200).json({foods: foods})
-        })
+        let filter = {}
+
+        // sets filter to only get approved foods or ones made by user if user is not admin
+        if (req.user.admin == false) {
+            filter = {
+                $or: [
+                    { godkendt: true },
+                    { user: req.user.id }
+                ]
+            }
+        }
+
+        // gets all foods with filter
+        const foods = await Food.find(filter);
+
+        // returns the foods
+        res.status(200).json({ foods: foods })
+
     } catch (error) {
         // return status 500 error if failed
-        console.log(error)
-        res.status(500).json({msg: "unable to get foods"})
+        res.status(500).json({ msg: "unable to get foods" })
     }
 }
 
-// exporting get method - gets specific food through id
+// gets specific food through barcode
 export const getFood = async (req, res) => {
     try {
-        const id = req.params.id
-        Food.findById(id)
-        .then((foods) => {
-            // returns status 200 and food that matches id
-            res.status(200).json({food: foods})
-        })
+        // finds food by barcode
+        const food = await Food.findOne({barcode: req.params.barcode})
+
+        // checks if user is admin or user who created the food
+        if (!(food.user == req.user.id || req.user.admin)) {
+            return res.status(401).json({ msg: "no access" })
+        }
+
+        // returns the food
+        res.status(200).json({ food })
+
     } catch (error) {
-        // logs and returns status 500 error if failed or no foods found
-        console.log(error)
-        res.status(500).json({msg: "unable to get food"})
+        // returns status 500 error if failed or no foods found
+        res.status(500).json({ msg: "unable to get food" })
     }
 };
 
@@ -37,7 +53,7 @@ export const search = async (req, res) => {
     try {
         // creates search term from the request query
         const searchTerm = req.query.searchTerm
-        
+
         // creates regex for search
         const searchRegex = new RegExp(searchTerm, "i")
         
@@ -64,32 +80,63 @@ export const search = async (req, res) => {
                 res.status(200).json({foods: [], msg: "no foods found"})
             }
         })
+
+        let filter = {
+            $or: [
+                { barcode: searchRegex },
+                { name: searchRegex }
+            ]
+        }
+
+        // sets filter to only get approved foods or ones made by user if user is not admin
+        if (req.user.admin == false) {
+            filter = {
+                $and: [
+                    { user: req.user.id },
+                    {
+                        $or: [
+                            { barcode: searchRegex },
+                            { name: searchRegex }
+                        ]
+                    }
+                ]
+            };
+        }
+
+        // searches for foods with the filter
+        const foods = await Food.find(filter)
+
+        // returns the foods
+        res.status(200).json({ foods })
+
     } catch (error) {
-        // logs and returns status 500 if error
-        console.log(error)
-        res.status(500).json({msg: "unable to get food"})
+        // returns status 500 if error
+        res.status(500).json({ msg: "unable to get food" })
     }
 };
 
 // exporting create method - creates new food
 export const createFood = async (req, res) => {
     try {
-        // creates a food from food body
-        const food = new Food(req.body)
+        // makes sure food gets user id from the user making the request
+        req.body.user = req.user.id
+
+        // makes godkendt false if user is not admin
+        if (req.user.admin == false) {
+            req.body.godkendt = false
+        }
 
         // saves the new food to database
-        await food.save()
-        .then((savedFoods) => {
-            // logs and returns the created food
-            console.log(savedFoods)
-            res.status(201).json({msg: 'food saved', food})
-        })
+        const food = await new Food(req.body).save()
+
+        // logs and returns the created food
+        res.status(201).json({ msg: 'food saved', food })
+
     } catch (error) {
-        // logs and returns status 500 if error -> food not created
-        console.log(error)
-        res.status(500).json({msg: 'unable to save new food'})
+        // returns status 500 if error -> food not created
+        res.status(500).json({ msg: 'unable to save new food' })
     }
-    
+
 }
 
 // exporting delete method - deletes a food through id
@@ -98,18 +145,27 @@ export const deleteFood = async (req, res) => {
         // gets id from params
         const id = req.params.id
 
-        // deletes the food from database
-        await Food.findByIdAndDelete(id)
-        .then((foods) => {
-            // returns the deleted food
-            res.status(200).json({msg: "Following food has been deleted", 
-                food: foods})
+        // finds food
+        const food = Food.findById(id)
+
+        // throws error if food is not found
+        if (!food) throw new Error("Food not found");
+
+        // 401 response if user is not admin or user who created the food
+        if (!(req.user.admin || req.user.id == food.user)) {
+            return res.status(401).json({ message: 'No access' });
+        }
+
+        const deletedFood = await Food.findByIdAndDelete(id);
+
+        // returns the deleted food
+        res.status(200).json({
+            msg: "Following food has been deleted",
+            food: deletedFood
         })
     } catch (error) {
-        // logs and returns status 500 if error 
-        // food not deleted or couldn't be found
-        console.log(error)
-        res.status(500).json({msg: "unable to delete food"})
+        // returns status 500 if error 
+        res.status(500).json({ msg: "unable to delete food" })
     }
 }
 
@@ -119,20 +175,13 @@ export const updateFood = async (req, res) => {
         // gets id from params
         const id = req.params.id
 
-        // sets new food with body  
-        const updatedFood = req.body
-
         // updates the food in database
-        await Food.findOneAndUpdate({_id: id}, updatedFood, {new: true})
-        .then((updatedFood) => {
-            // logs and return the updated food
-            console.log(updatedFood)
-            res.status(200).json({msg: "food updated", food: updatedFood})
-        })
+        const updatedFood = await Food.findOneAndUpdate({ _id: id }, req.body, { new: true })
+
+        // return the updated food
+        res.status(200).json({ msg: "food updated", food: updatedFood })
     } catch (error) {
-        // logs and returns status 500 if error 
-        // food couldn't be found or couldn't be updated
-        console.log(error)
-        res.status(500).json({msg: "unable to update food"})
+        // returns status 500 if error 
+        res.status(500).json({ msg: "unable to update food" })
     }
 }
